PLT-4858: Allow system admin to delete all channels. (#5180)
The app-package refactor means that the check for a Channel Member
object before allowing to delete the channel is now redundant, as the
check is already carried out by using HasPermissionToChannel() on the
user requesting it. As a result, System Admins can now delete channels
through the API even if they aren't members.